backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis[asyncio]==5.2.1

# Scheduler
apscheduler==3.10.4

# HTTP client
httpx==0.27.2

# Auth
PyJWT==2.10.1
bcrypt==5.0.0

# Validation & settings
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20

